eshop/requirements.txt
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
colorama==0.4.6
Django==4.2.23
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
exceptiongroup==1.3.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
redis==3.5.3
redis-py-cluster==2.1.3
referencing==0.36.2
rpds-py==0.25.1
sentry-sdk==2.29.1
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
upstash-redis==1.4.0
uritemplate==4.2.0
urllib3==2.4.0